# Conflicts:
#	src/main/java/com/ii/testautomation/controllers/SubModulesController.java
#	src/main/java/com/ii/testautomation/repositories/MainModulesRepository.java
#	src/main/java/com/ii/testautomation/repositories/TestTypesRepository.java
#	src/main/java/com/ii/testautomation/service/MainModulesService.java
#	src/main/java/com/ii/testautomation/service/ProjectService.java
#	src/main/java/com/ii/testautomation/service/TestTypesService.java
#	src/main/java/com/ii/testautomation/service/impl/MainModulesServiceImpl.java
#	src/main/java/com/ii/testautomation/service/impl/ProjectServiceImpl.java
#	src/main/java/com/ii/testautomation/service/impl/TestTypesServiceImpl.java
#	src/main/java/com/ii/testautomation/utils/Constants.java
#	src/main/java/com/ii/testautomation/utils/EndpointURI.java
#	src/main/java/com/ii/testautomation/utils/StatusCodeBundle.java
#	src/main/resources/MessagesAndCodes.properties

+++ src/main/java/com/ii/testautomation/controllers/SubModulesController.java
@@ -1,8 +1,6 @@
 package com.ii.testautomation.controllers;
 
 import com.ii.testautomation.dto.request.SubModulesRequest;
-import com.ii.testautomation.dto.response.SubModulesResponse;
-import com.ii.testautomation.dto.search.ProjectSearch;
 import com.ii.testautomation.dto.search.SubModuleSearch;
 import com.ii.testautomation.enums.RequestStatus;
 import com.ii.testautomation.response.common.BaseResponse;
@@ -10,6 +8,7 @@ import com.ii.testautomation.response.common.ContentResponse;
 import com.ii.testautomation.response.common.PaginatedContentResponse;
 import com.ii.testautomation.service.MainModulesService;
 import com.ii.testautomation.service.SubModulesService;
+import com.ii.testautomation.service.TestCasesService;
 import com.ii.testautomation.utils.Constants;
 import com.ii.testautomation.utils.EndpointURI;
 import com.ii.testautomation.utils.StatusCodeBundle;
@@ -30,27 +29,24 @@ public class SubModulesController {
     private SubModulesService subModulesService;
     @Autowired
     private MainModulesService mainModulesService;
-
+    @Autowired
+    private TestCasesService testCasesService;
     @Autowired
     private StatusCodeBundle statusCodeBundle;
 
     @PostMapping(value = EndpointURI.SUBMODULE)
-    public ResponseEntity<Object> saveSubModules(@RequestBody SubModulesRequest subModulesRequest)
-    {
-        if(subModulesService.existsBySubModulesName(subModulesRequest.getName()))
-        {
+    public ResponseEntity<Object> saveSubModules(@RequestBody SubModulesRequest subModulesRequest) {
+        if (subModulesService.existsBySubModulesName(subModulesRequest.getName())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesAlReadyExistCode(),
                     statusCodeBundle.getSubModuleNameAlReadyExistMessage()));
         }
-        if(subModulesService.existsBySubModulesPrefix(subModulesRequest.getPrefix()))
-        {
+        if (subModulesService.existsBySubModulesPrefix(subModulesRequest.getPrefix())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesAlReadyExistCode(),
                     statusCodeBundle.getSubModulePrefixAlReadyExistMessage()));
         }
-        if (!mainModulesService.isExistModulesId(subModulesRequest.getMain_module_Id()))
-        {
+        if (!mainModulesService.existsByMainModuleId(subModulesRequest.getMain_module_Id())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getMainModulesNotExistCode(),
                     statusCodeBundle.getMainModuleNotExistsMessage()));
@@ -61,45 +57,38 @@ public class SubModulesController {
                 statusCodeBundle.getSaveSubModuleSuccessMessage()));
 
     }
-    @PutMapping(value=EndpointURI.SUBMODULE)
-    public ResponseEntity<Object> editSubModules(@RequestBody SubModulesRequest subModulesRequest)
-    {
-        if (!subModulesService.existsBySubModuleId(subModulesRequest.getId()))
-        {
+
+    @PutMapping(value = EndpointURI.SUBMODULE)
+    public ResponseEntity<Object> editSubModules(@RequestBody SubModulesRequest subModulesRequest) {
+        if (!subModulesService.existsBySubModuleId(subModulesRequest.getId())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesNotExistCode(),
                     statusCodeBundle.getSubModuleNotExistsMessage()));
         }
-        if (!mainModulesService.isExistMainModulesId(subModulesRequest.getMain_module_Id()))
-        {
+        if (!mainModulesService.existsByMainModuleId(subModulesRequest.getMain_module_Id())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getMainModulesNotExistCode(),
                     statusCodeBundle.getMainModuleNotExistsMessage()));
         }
-        if(subModulesService.isUpdateSubModuleNameExits(subModulesRequest.getName(),subModulesRequest.getId()))
-        {
+        if (subModulesService.isUpdateSubModuleNameExits(subModulesRequest.getName(), subModulesRequest.getId())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesAlReadyExistCode(),
                     statusCodeBundle.getSubModuleNameAlReadyExistMessage()));
         }
-        if(subModulesService.isUpdateSubModulePrefixExits(subModulesRequest.getPrefix(),subModulesRequest.getId()))
-        {
+        if (subModulesService.isUpdateSubModulePrefixExits(subModulesRequest.getPrefix(), subModulesRequest.getId())) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesAlReadyExistCode(),
                     statusCodeBundle.getSubModulePrefixAlReadyExistMessage()));
         }
-
         subModulesService.saveSubModules(subModulesRequest);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getUpdateSubModuleSuccessMessage()));
-
     }
+
     @GetMapping(value = EndpointURI.SUBMODULE_BY_ID)
-    public ResponseEntity<Object> getSubModuleById(@PathVariable Long id)
-    {
-        if (!subModulesService.existsBySubModuleId(id))
-        {
+    public ResponseEntity<Object> getSubModuleById(@PathVariable Long id) {
+        if (!subModulesService.existsBySubModuleId(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesNotExistCode(),
                     statusCodeBundle.getSubModuleNotExistsMessage()));
@@ -111,28 +100,24 @@ public class SubModulesController {
     }
 
     @GetMapping(value = EndpointURI.SUBMODULE_BY_MAIN_MODULE_ID)
-    public ResponseEntity<Object> getSubModuleByMainModuleId(@PathVariable Long id)
-    {
-        if (!mainModulesService.isExistMainModulesId(id))
-        {
+    public ResponseEntity<Object> getSubModuleByMainModuleId(@PathVariable Long id) {
+        if (!mainModulesService.existsByMainModuleId(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getMainModulesNotExistCode(),
                     statusCodeBundle.getMainModuleNotExistsMessage()));
         }
-        List<SubModulesResponse> subModulesResponseList=subModulesService.getAllSubModuleByMainModuleId(id);
-        if(subModulesResponseList.isEmpty())
-        {
+        if (!subModulesService.existsByMainModuleId(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getFailureCode(),
                     statusCodeBundle.getGetSubModuleNotHaveMainModuleId()));
         }
-
         return ResponseEntity.ok(new ContentResponse<>(Constants.SUBMODULES,
                 subModulesService.getAllSubModuleByMainModuleId(id),
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getGetSubModulesSuccessMessage()));
     }
+
     @GetMapping(value = EndpointURI.SUBMODULES_SEARCH)
     public ResponseEntity<Object> getALlSubModuleWithMultiSearch(@RequestParam(name = "page") int page,
                                                                  @RequestParam(name = "size") int size,
@@ -141,24 +126,27 @@ public class SubModulesController {
                                                                  SubModuleSearch subModuleSearch) {
         Pageable pageable = PageRequest.of(page, size, Sort.Direction.valueOf(direction), sortField);
         PaginatedContentResponse.Pagination pagination = new PaginatedContentResponse.Pagination(page, size, 0, 0l);
-        return ResponseEntity.ok(new ContentResponse<>(Constants.SUBMODULES, subModulesService.multiSearchSubModule(pageable,pagination,subModuleSearch),
+        return ResponseEntity.ok(new ContentResponse<>(Constants.SUBMODULES, subModulesService.multiSearchSubModule(pageable, pagination, subModuleSearch),
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getGetAllSubModuleSuccessMessage()));
     }
+
     @DeleteMapping(value = EndpointURI.SUBMODULE_BY_ID)
-    public ResponseEntity<Object> deleteSubModuleById(@PathVariable Long id)
-    {
-        if (!subModulesService.existsBySubModuleId(id))
-        {
+    public ResponseEntity<Object> deleteSubModuleById(@PathVariable Long id) {
+        if (!subModulesService.existsBySubModuleId(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getSubModulesNotExistCode(),
                     statusCodeBundle.getSubModuleNotExistsMessage()));
+        }
+        if (testCasesService.existsBySubModuleId(id)) {
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getSubModulesDependentCode(),
+                    statusCodeBundle.getSubModulesDependentMessage()));
         }
         subModulesService.deleteSubModuleById(id);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getDeleteSubModuleSuccessMessage()));
     }
-
 }
